Threading, Plugin system rewrite begins
