+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
     "LineNumberDeco.enableRainbow": false,
-    "CodeGPT.apiKey": "Ollama"
+    "CodeGPT.apiKey": "Ollama",
+    "LineNumberDeco.enableRelativeLine": false
 }
