+++ Cargo.toml
@@ -4,51 +4,19 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-axum = { version = "0.7.5", features = ["tokio", "http1"] }
-console_log = "1.0.0"
-hyper = "1.3.1"
-reqwest = "0.12.5"
-serde = "1.0.210"
+serde = "1.0.214"
 serde_json = "1.0.128"
-socketioxide = "0.15.0"
+socketioxide = "0.15.1"
 tokio = { version = "1.37.0", features = ["rt", "net", "rt-multi-thread"] }
 tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-viz = "0.8.4"
-bincode = "1.3.3"
-ansi_term = "0.12.1"
-monoio = { version = "0.2", features = ["sync"] }
-tokio-console = "0.1.11"
-console-subscriber = "0.3.0"
-PebbleVault = { git = "https://github.com/Far-Beyond-Dev/PebbleVault.git" }
-bindgen = "0.69.4"
-http = "1.1.0"
-bytes = "1.6.0"
-http-body-util = "0.1.2"
-
-
+viz = "0.9.0"
+uuid = "1.10.0"
+libloading = "0.8.5"
+notify = "7.0.0"
+horizon_data_types = "0.3.0"
 plugin-test-api = { version = "^0.1", path = "./plugin-api" }
 plugin-test-plugins = { version = "^0.1", path = "./plugins" }
-
-nalgebra = "0.33.0"
-v = "0.1.0"
 rayon = "1.10.0"
-log = "0.4.22"
-csv = "1.3.0"
-uid = "0.1.7"
-uuid = "1.10.0"
-env_logger = "0.11.5"
-chrono = "0.4.38"
-libloading = "0.8.5"
-notify = "6.1.1"
-notify-debouncer-full = "0.3.1"
-horizon_data_types = "0.3.0"
-colored = "2.0.0"
-
-rand = "0.8.5"
-lazy_static = "1.5.0"
-ez_logging = "0.1.1"
-fmt = "0.1.0"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 mimalloc = "0.1.43"
@@ -68,3 +36,10 @@ strip = true
 inherits = "release"
 debug = true
 strip = false
+
+
+
+[profile.release.package."*"]
+opt-level = 3
+strip = true
+codegen-units = 1
