+++ plugin-api/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 
 [dependencies]
 async-trait = "0.1.83"
-ez_logging = "0.1.1"
 horizon_data_types = "0.3.0"
 tokio = { version = "1.37.0", features = ["rt", "net", "rt-multi-thread"] }
 uuid = "1.11.0"
