+++ plugins/Cargo.toml
@@ -9,13 +9,10 @@ edition = "2021"
 
 # generated file, do not edit manually
 [dependencies]
-nom = "7.1.3"
 async-trait = "0.1.50" 
 horizon_data_types = "0.3.0"
 tokio = { version = "1.37.0", features = ["rt", "net", "rt-multi-thread"] }
-rand = "0.8.5"
 socketioxide = "0.15.0"
-ez_logging = "0.1.1"
 serde = "1.0.210"
 uuid = "1.10.0"
 serde_json = "1.0.128"
