+++ plugins/test_plugin/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "test_plugin"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
